Added a start to the file system
There is a File Manager class that will deal with all the additions and
deletions from the stadium information as well as the file persistence
wile the program is up and running.

I've commented the methods nicely so it shouldn't be to hard to figure
out what they do but I will explain everything when we meet on Tuesday.

Also, I added a text file of how my code manages the data. We can also
talk about this on Tuesday if there are any issues.