Added Souvenir Class Folder
I've added a file containing a text file, header file and  cpp file for
dealing with all the souvenirs within the project.
There are methods for updating item prices, deleting items and adding
items. All of these methods are tested and - at least for testing that I
did - are working.

There is a method updateSouvenirList() for keeping data persistent
throughout the runs. Run that method and all the changes that have been
made to the souvenir class will be made permanent within the text file.

PLEASE NOTE: whom ever uses this class will need to change the text from
normal C++ file implementation to the QT file implementation as I don't
want to implement it incorrectly.

I'll send out an email with the same information so we're all on the
same page.