Prims now functional
Prims returns a queue of edges that are included in the MST as well as
the weight of the MST (Passed back by reference). I put an instance of
the filemanager in the PrimWindow to allow it to work without clicking
on the Visit Stadiums/Dream Vacation buttons. The correct MST weight is
displayed already.