Idea for a simple solution to the date problem.
I could modify the the stadium information to include a simple date
class that can be used to organize stadiums in chronological order, but
I was thinking if this would be an easier (maybe less buggy) solution.
Would you rather me simply keep the date-opened element as a string but
changing the format from what it was to something like YYYY/MM/DD?
That way it would still compare strings like with Ahmad's other ways of
organizing data.

Example: if one stadium opened in 1977/12/01 and another one opened in
1978/4/5, the 1977 stadium would still be in sorted order by their
strings.

I'll send an email about this, too.